LuceneOptimizedWrappedDirectory needs to allow for extension #2180
